+++ src/components/MenuListPage/MenuPreview.tsx
@@ -3,7 +3,7 @@ import imagePlaceholder from "../../resources/image-placeholder.png";
 import { formatPrice } from "../../lib/formatting";
 import styles from "./MenuPreview.module.css";
 import { Link } from "react-router-dom";
-import { Menu } from "../../types/types";
+import { Menu } from "../../lib/types";
 
 export default function MenuPreview({
   menu,
